Avoid creating multiple clients per project
If a file is opened within a project while a client for that
project already exists, we create a new client to replace
the old one in the map, without disposing of the old client.

Fixes:
1. If createClient replaces a client in the map, always dispose
of the replaced client to avoid leaking clients.
2. Remove now redundant calls to disposeClient which are made
just before calls to createClient for the same project.
3. Don't call createClient at all when a document is opened
in a project if a client already exists for the given project.